Build ParticipantMixin with common code for ParticipantList and ParticipantDetail views.
